Add decimal support by updating inputs to step="any" and switching parsing from parseInt to parseFloat in multiple components (AgentsManager, CommunityManagement, DamageModal, InventoryImportModal, PayrollRunModal, ProductModal, ProductVariantsModal, ReturnModal, SalesRecorder, StockTransferModal, VariantSelectorModal). Also adjusted related data parsing to use floating values.

X-Lovable-Edit-ID: edt-45807fd6-3710-47e9-aa5d-e7fd06a077f7

+++ src/components/dashboard/CommunityManagement.tsx
@@ -127,7 +127,7 @@ export function CommunityManagement() {
     const { error } = await supabase.from("wash_forums").insert({
       name: newCommunity.name,
       province: newCommunity.province,
-      community_size: parseInt(newCommunity.community_size) || 0,
+      community_size: parseFloat(newCommunity.community_size) || 0,
       description: newCommunity.description,
       products_needed: newCommunity.products_needed,
       priority: newCommunity.priority,
